etl/postgres_to_es/requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
pydantic==2.6.0
pydantic-settings==2.1.0
redis==5.0.1
PyYAML==6.0.1
